+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Change import of `exportToCsv` from `stripes-util` to `stripes-components`. Refs UIU-3202.
 * Provide `itemToString` to create unique `key` attributes. Refs UIU-3312.
 * Return appropriate error message when item for manual fee/fine can't be found. Refs UIU-2701.
+* Add `pronouns` field to user edit form. Refs UIU-3119.
 
 ## [11.0.11](https://github.com/folio-org/ui-users/tree/v11.0.11) (2025-01-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v11.0.10...v11.0.11)